another refactoring of message construction

+++ src/main/java/com/greenfox/p2pchat/repositories/MessageRepository.java
@@ -3,5 +3,5 @@ package com.greenfox.p2pchat.repositories;
 import com.greenfox.p2pchat.models.Message;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageRepository extends CrudRepository<Message, Long> {
+public interface MessageRepository extends CrudRepository<Message, Integer> {
 }
